requirements.txt
# FastAPI and Uvicorn
fastapi==0.124.4
uvicorn[standard]==0.38.0

# Kite Connect API
kiteconnect==5.0.1

# WebSocket and other client libraries
websocket-client==1.3.2

# Data Analysis and Machine Learning
pandas==1.5.0
numpy==1.23.4
scikit-learn==1.1.2
nltk==3.7

# HTTP Requests
requests==2.32.5

# Environment Variables and Secrets Management
python-dotenv==0.19.2

# SQLite and Redis support
sqlitedict==1.7.0
redis==4.0.0

# Stock and Financial Data Libraries
yfinance==0.1.75
alpha_vantage==2.3.1

# Database support
SQLAlchemy==1.4.40
psycopg2-binary==2.9.3

# AWS S3 support
boto3==1.24.9

# For background task scheduling
APScheduler==3.9.1

# Security Libraries
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
authlib==1.0.0

# Asynchronous HTTP requests
httpx==0.23.0

# Form and file handling
python-multipart==0.0.5

# Alembic for database migrations
alembic==1.8.1


setuptools>=42.0.0
